requirements.txt
# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Frontend
streamlit==1.28.1

# LLM and AI
openai==1.30.5
langchain==0.0.350
langchain-openai==0.0.2

# Template engine for code generation
jinja2==3.1.2

# Environment variables
python-dotenv==1.0.0

# HTTP requests
httpx==0.25.2

# Data handling
pydantic==2.5.0

# File operations
python-multipart==0.0.6

# RAG and ML dependencies
scikit-learn==1.3.2
numpy==1.24.3

# Development tools
pytest==7.4.3
black==23.11.0 